+++ solcx/exceptions.py
@@ -1,43 +1,37 @@
-import textwrap
-
-from .utils.string import force_text
-
-
-def force_text_maybe(value, encoding="iso-8859-1"):
-    if value is not None:
-        return force_text(value)
-
-
-DEFAULT_MESSAGE = "An error occurred during execution"
+from typing import Dict, List
 
 
 class SolcError(Exception):
-    message = DEFAULT_MESSAGE
+    message = "An error occurred during execution"
 
-    def __init__(self, command, return_code, stdin_data, stdout_data, stderr_data, message=None):
+    def __init__(
+        self,
+        message: str = None,
+        command: List = None,
+        return_code: int = None,
+        stdin_data: str = None,
+        stdout_data: str = None,
+        stderr_data: str = None,
+        error_dict: Dict = None,
+    ) -> None:
         if message is not None:
             self.message = message
-        self.command = command
+        self.command = command or []
         self.return_code = return_code
-        self.stdin_data = force_text_maybe(stdin_data, "utf8")
-        self.stderr_data = force_text_maybe(stderr_data, "utf8")
-        self.stdout_data = force_text_maybe(stdout_data, "utf8")
+        self.stdin_data = stdin_data
+        self.stderr_data = stderr_data
+        self.stdout_data = stdout_data
+        self.error_dict = error_dict
 
-    def __str__(self):
-        return textwrap.dedent(
-            (
-                """
-        {s.message}
-        > command: `{command}`
-        > return code: `{s.return_code}`
-        > stderr:
-        {s.stdout_data}
-        > stdout:
-        {s.stderr_data}
-        """
-            ).format(
-                s=self, command=" ".join(self.command),
-            )
+    def __str__(self) -> str:
+        return (
+            f"{self.message}"
+            f"\n> command: `{' '.join(str(i) for i in self.command)}`"
+            f"\n> return code: `{self.return_code}`"
+            "\n> stdout:"
+            f"\n{self.stdout_data}"
+            "\n> stderr:"
+            f"\n{self.stderr_data}"
         ).strip()
 
 
@@ -45,9 +39,33 @@ class ContractsNotFound(SolcError):
     message = "No contracts found during compilation"
 
 
+class SolcInstallationError(Exception):
+    pass
+
+
+class UnknownOption(AttributeError):
+    pass
+
+
+class UnknownValue(ValueError):
+    pass
+
+
+class UnexpectedVersionError(Exception):
+    pass
+
+
+class UnsupportedVersionError(ValueError):
+    pass
+
+
 class SolcNotInstalled(Exception):
     pass
 
 
 class DownloadError(Exception):
+    pass
+
+
+class UnexpectedVersionWarning(Warning):
     pass