+++ solcx/utils/lock.py
@@ -3,7 +3,7 @@ import sys
 import tempfile
 import threading
 from pathlib import Path
-from typing import Dict
+from typing import Any, Dict, Union
 
 if sys.platform == "win32":
     import msvcrt
@@ -15,11 +15,11 @@ else:
     NON_BLOCKING = fcntl.LOCK_EX | fcntl.LOCK_NB
     BLOCKING = fcntl.LOCK_EX
 
-_locks: Dict = {}
+_locks: Dict[str, Union["UnixLock", "WindowsLock"]] = {}
 _base_lock = threading.Lock()
 
 
-def get_process_lock(lock_id):
+def get_process_lock(lock_id: str) -> Union["UnixLock", "WindowsLock"]:
     with _base_lock:
         if lock_id not in _locks:
             if sys.platform == "win32":
@@ -30,18 +30,24 @@ def get_process_lock(lock_id):
 
 
 class _ProcessLock:
-    def __init__(self, lock_id):
+    """
+    Ensure an action is both thread-safe and process-safe.
+    """
+
+    def __init__(self, lock_id: str) -> None:
         self._lock = threading.Lock()
         self._lock_path = Path(tempfile.gettempdir()).joinpath(f".solcx-lock-{lock_id}")
         self._lock_file = self._lock_path.open("w")
 
-    def wait(self):
-        self.acquire(True)
-        self.release()
-
 
 class UnixLock(_ProcessLock):
-    def acquire(self, blocking):
+    def __enter__(self) -> None:
+        self.acquire(True)
+
+    def __exit__(self, *args: Any) -> None:
+        self.release()
+
+    def acquire(self, blocking: bool) -> bool:
         if not self._lock.acquire(blocking):
             return False
         try:
@@ -51,19 +57,27 @@ class UnixLock(_ProcessLock):
             return False
         return True
 
-    def release(self):
+    def release(self) -> None:
         fcntl.flock(self._lock_file, fcntl.LOCK_UN)
         self._lock.release()
 
 
 class WindowsLock(_ProcessLock):
-    def acquire(self, blocking):
+    def __enter__(self) -> None:
+        self.acquire(True)
+
+    def __exit__(self, *args: Any) -> None:
+        self.release()
+
+    def acquire(self, blocking: bool) -> bool:
         if not self._lock.acquire(blocking):
             return False
         while True:
             try:
-                fd = os.open(self._lock_path, OPEN_MODE)
-                msvcrt.locking(fd, msvcrt.LK_LOCK if blocking else msvcrt.LK_NBLCK, 1)
+                fd = os.open(self._lock_path, OPEN_MODE)  # type: ignore
+                msvcrt.locking(  # type: ignore
+                    fd, msvcrt.LK_LOCK if blocking else msvcrt.LK_NBLCK, 1  # type: ignore
+                )
                 self._fd = fd
                 return True
             except OSError:
@@ -71,6 +85,6 @@ class WindowsLock(_ProcessLock):
                     self._lock.release()
                     return False
 
-    def release(self):
-        msvcrt.locking(self._fd, msvcrt.LK_UNLCK, 1)
+    def release(self) -> None:
+        msvcrt.locking(self._fd, msvcrt.LK_UNLCK, 1)  # type: ignore
         self._lock.release()