+++ solcx/wrapper.py
@@ -1,189 +1,149 @@
-from __future__ import absolute_import
-
+import re
 import subprocess
+from pathlib import Path
+from typing import Any, Dict, List, Tuple, Union
 
 from semantic_version import Version
 
-from .exceptions import SolcError
-from .install import get_executable
-from .utils.string import coerce_return_to_text, force_bytes
+from solcx import install
+from solcx.exceptions import SolcError, UnknownOption, UnknownValue
 
 
-@coerce_return_to_text
+def _get_solc_version(solc_binary: Union[Path, str]) -> Version:
+    # private wrapper function to get `solc` version
+    stdout_data = subprocess.check_output([solc_binary, "--version"], encoding="utf8")
+    version_str = re.findall(r"(?<=Version: ).*?(?=\+)", stdout_data)[0]
+    version_str = re.sub(r"\.0(?=[1-9])", ".", version_str)
+    return Version.coerce(version_str)
+
+
+def _to_string(key: str, value: Any) -> str:
+    # convert data into a string prior to calling `solc`
+    if isinstance(value, (int, str)):
+        return str(value)
+    elif isinstance(value, Path):
+        return value.as_posix()
+    elif isinstance(value, (list, tuple)):
+        return ",".join(_to_string(key, i) for i in value)
+    else:
+        raise TypeError(f"Invalid type for {key}: {type(value)}")
+
+
 def solc_wrapper(
-    solc_binary=None,
-    stdin=None,
-    help=None,
-    version=None,
-    add_std=None,
-    combined_json=None,
-    optimize=None,
-    optimize_runs=None,
-    libraries=None,
-    output_dir=None,
-    gas=None,
-    assemble=None,
-    link=None,
-    source_files=None,
-    import_remappings=None,
-    ast=None,
-    ast_json=None,
-    asm=None,
-    asm_json=None,
-    opcodes=None,
-    bin=None,
-    bin_runtime=None,
-    clone_bin=None,
-    abi=None,
-    hashes=None,
-    userdoc=None,
-    devdoc=None,
-    formal=None,
-    allow_paths=None,
-    base_path=None,
-    standard_json=None,
-    success_return_code=None,
-    evm_version=None,
-):
-    if solc_binary is None:
-        solc_binary = get_executable()
+    solc_binary: Union[Path, str] = None,
+    stdin: str = None,
+    source_files: Union[List, Path, str] = None,
+    import_remappings: Union[Dict, List, str] = None,
+    success_return_code: int = None,
+    **kwargs: Any,
+) -> Tuple[str, str, List, subprocess.Popen]:
+    """
+    Wrapper function for calling to `solc`.
 
-    command = [solc_binary]
+    Arguments
+    ---------
+    solc_binary : Path | str, optional
+        Location of the `solc` binary. If not given, the current default binary is used.
+    stdin : str, optional
+        Input to pass to `solc` via stdin
+    source_files : list | Path | str, optional
+        Path, or list of paths, of sources to compile
+    import_remappings : Dict | List | str,  optional
+        Path remappings. May be given as a string or list of strings formatted as `"prefix=path"`
+        or a dict of `{"prefix": "path"}`
+    success_return_code : int, optional
+        Expected exit code. Raises `SolcError` if the process returns a different value.
 
-    solc_version = Version(solc_binary.rsplit("-v")[-1].split("\\")[0])
-    solc_minor = solc_version.minor
+    Keyword Arguments
+    -----------------
+    **kwargs : Any
+        Flags to be passed to `solc`. Keywords are converted to flags by prepending `--` and
+        replacing `_` with `-`, for example the keyword `evm_version` becomes `--evm-version`.
+        Values may be given in the following formats:
 
-    if help:
-        command.append("--help")
-        if success_return_code is None:
-            success_return_code = 1
-    elif success_return_code is None:
-        success_return_code = 0
+            * `False`, `None`: ignored
+            * `True`: flag is used without any arguments
+            * str: given as an argument without modification
+            * int: given as an argument, converted to a string
+            * Path: converted to a string via `Path.as_posix()`
+            * List, Tuple: elements are converted to strings and joined with `,`
 
-    if version:
-        command.append("--version")
+    Returns
+    -------
+    str
+        Process `stdout` output
+    str
+        Process `stderr` output
+    List
+        Full command executed by the function
+    Popen
+        Subprocess object used to call `solc`
+    """
+    if solc_binary:
+        solc_binary = Path(solc_binary)
+    else:
+        solc_binary = install.get_executable()
 
-    # removed in 0.4.21 and does nothing since <0.4.11, should be removed in the future
-    if add_std:
-        command.append("--add-std")
+    solc_version = _get_solc_version(solc_binary)
+    command: List = [solc_binary]
 
-    if optimize:
-        command.append("--optimize")
-
-    if optimize_runs is not None:
-        command.extend(("--optimize-runs", str(optimize_runs)))
-
-    if link:
-        command.append("--link")
-
-    if libraries is not None:
-        command.extend(("--libraries", libraries))
-
-    if output_dir is not None:
-        command.extend(("--output-dir", output_dir))
-
-    if combined_json:
-        if solc_minor >= 5:
-            combined_json = combined_json.replace(",clone-bin", "")
-        command.extend(("--combined-json", combined_json))
-
-    if gas:
-        command.append("--gas")
-
-    if allow_paths:
-        command.extend(("--allow-paths", allow_paths))
-
-    if standard_json:
-        command.append("--standard-json")
-
-    if assemble:
-        command.append("--assemble")
-
-    if import_remappings is not None:
-        command.extend(import_remappings)
+    if success_return_code is None:
+        success_return_code = 1 if "help" in kwargs else 0
 
     if source_files is not None:
-        command.extend(source_files)
+        if isinstance(source_files, (str, Path)):
+            command.append(_to_string("source_files", source_files))
+        else:
+            command.extend([_to_string("source_files", i) for i in source_files])
 
-    # Output configuration
-    if ast_json:
-        command.append("--ast-json")
+    if import_remappings is not None:
+        if isinstance(import_remappings, str):
+            command.append(import_remappings)
+        else:
+            if isinstance(import_remappings, dict):
+                import_remappings = [f"{k}={v}" for k, v in import_remappings.items()]
+            command.extend(import_remappings)
 
-    if asm:
-        command.append("--asm")
+    for key, value in kwargs.items():
+        if value is None or value is False:
+            continue
 
-    if asm_json:
-        command.append("--asm-json")
+        key = f"--{key.replace('_', '-')}"
+        if value is True:
+            command.append(key)
+        else:
+            command.extend([key, _to_string(key, value)])
 
-    if opcodes:
-        command.append("--opcodes")
-
-    if bin:
-        command.append("--bin")
-
-    if bin_runtime:
-        command.append("--bin-runtime")
-
-    if abi:
-        command.append("--abi")
-
-    if hashes:
-        command.append("--hashes")
-
-    if userdoc:
-        command.append("--userdoc")
-
-    if devdoc:
-        command.append("--devdoc")
-
-    if evm_version:
-        command.extend(("--evm-version", evm_version))
-
-    # unsupported by <0.6.9
-    if base_path:
-        if solc_version <= Version("0.6.8"):
-            raise AttributeError(
-                "solc {} does not support the --base-path flag".format(solc_version)
-            )
-        command.extend(("--base-path", base_path))
-
-    # unsupported by >=0.6.0
-    if ast:
-        if solc_minor >= 6:
-            raise AttributeError("solc 0.{}.x does not support the --ast flag".format(solc_minor))
-        command.append("--ast")
-
-    # unsupported by >=0.5.0
-    if clone_bin:
-        if solc_minor >= 5:
-            raise AttributeError(
-                "solc 0.{}.x does not support the --clone-bin flag".format(solc_minor)
-            )
-        command.append("--clone-bin")
-
-    if formal:
-        if solc_minor >= 5:
-            raise AttributeError(
-                "solc 0.{}.x does not support the --formal flag".format(solc_minor)
-            )
-        command.append("--formal")
-
-    if not standard_json and not source_files:
+    if "standard_json" not in kwargs and not source_files:
         # indicates that solc should read from stdin
         command.append("-")
 
     if stdin is not None:
-        # solc seems to expects utf-8 from stdin:
-        # see Scanner class in Solidity source
-        stdin = force_bytes(stdin, "utf8")
+        stdin = str(stdin)
 
     proc = subprocess.Popen(
-        command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+        command,
+        stdin=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        encoding="utf8",
     )
 
     stdoutdata, stderrdata = proc.communicate(stdin)
 
     if proc.returncode != success_return_code:
+        if stderrdata.startswith("unrecognised option"):
+            # unrecognised option '<FLAG>'
+            flag = stderrdata.split("'")[1]
+            raise UnknownOption(f"solc {solc_version} does not support the '{flag}' option'")
+        if stderrdata.startswith("Invalid option"):
+            # Invalid option to <FLAG>: <OPTION>
+            flag, option = stderrdata.split(": ")
+            flag = flag.split(" ")[-1]
+            raise UnknownValue(
+                f"solc {solc_version} does not accept '{option}' as an option for the '{flag}' flag"
+            )
+
         raise SolcError(
             command=command,
             return_code=proc.returncode,