+++ tests/install/test_locks.py
@@ -3,6 +3,8 @@
 import multiprocessing as mp
 import threading
 
+import pytest
+
 import solcx
 
 
@@ -24,19 +26,21 @@ class ThreadWrap:
             raise self.exc
 
 
+@pytest.mark.skipif("'--no-install' in sys.argv")
 def test_threadlock(nosolc):
-    threads = [ThreadWrap(solcx.install_solc, "0.5.0") for i in range(4)]
+    threads = [ThreadWrap(solcx.install_solc, "0.6.9") for i in range(4)]
     for t in threads:
         t.join()
 
 
+@pytest.mark.skipif("'--no-install' in sys.argv")
 def test_processlock(nosolc):
     # have to use a context here to prevent a conflict with tqdm
     ctx = mp.get_context("spawn")
-    threads = [ctx.Process(target=solcx.install_solc, args=("0.5.0",),) for i in range(4)]
+    threads = [ctx.Process(target=solcx.install_solc, args=("0.6.9",)) for i in range(4)]
     for t in threads:
         t.start()
-    solcx.install_solc("0.5.0")
+    solcx.install_solc("0.6.9")
     for t in threads:
         t.join()
         assert t.exitcode == 0